Working rects

+++ plugins/canvas/src/LinearBasicDisplay/components/hitTesting.test.ts
@@ -59,7 +59,7 @@ function makeData(
     rectPositions: new Uint32Array(0),
     rectYs: new Float32Array(0),
     rectHeights: new Float32Array(0),
-    rectColors: new Uint8Array(0),
+    rectColors: new Uint32Array(0),
     rectFeatureIndices: new Uint32Array(0),
     linePositions: new Uint32Array(0),
     lineYs: new Float32Array(0),
