+++ plugins/canvas/src/LinearBasicDisplay/fetchAutorun.test.ts
@@ -26,7 +26,7 @@ function makeEmptyFeatureData(regionStart: number): FeatureDataResult {
     rectPositions: new Uint32Array(0),
     rectYs: new Float32Array(0),
     rectHeights: new Float32Array(0),
-    rectColors: new Uint8Array(0),
+    rectColors: new Uint32Array(0),
     rectFeatureIndices: new Uint32Array(0),
     linePositions: new Uint32Array(0),
     lineYs: new Float32Array(0),
